feat: Unify the different type definitions of Config
