+++ crates/cac_client/Cargo.toml
@@ -14,7 +14,6 @@ jsonlogic = { workspace = true }
 log = { workspace = true }
 once_cell = { workspace = true }
 reqwest = { workspace = true }
-serde = { workspace = true }
 serde_json = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
