+++ crates/context_aware_config/src/api/config.rs
@@ -1,4 +1,3 @@
 mod handlers;
-pub mod types;
 pub use handlers::endpoints;
 mod helpers;
