+++ crates/context_aware_config/src/helpers.rs
@@ -1,17 +1,4 @@
-use crate::{
-    api::config::types::{Config, Context},
-    db::{
-        models::ConfigVersion,
-        schema::{
-            config_versions,
-            contexts::dsl::{self as ctxt},
-            default_configs::dsl as def_conf,
-        },
-    },
-};
-
-#[cfg(feature = "high-performance-mode")]
-use crate::db::schema::event_log::dsl as event_log;
+use std::collections::HashMap;
 
 use actix_web::http::header::{HeaderMap, HeaderName, HeaderValue};
 use actix_web::web::Data;
@@ -33,13 +20,23 @@ use service_utils::{
     helpers::{generate_snowflake_id, validation_err_to_str},
     service::types::AppState,
 };
-
 use superposition_macros::{db_error, unexpected_error, validation_error};
-use superposition_types::{result as superposition, Cac, Condition, Overrides};
+use superposition_types::{
+    result as superposition, Cac, Condition, Config, Context, Overrides,
+};
 #[cfg(feature = "high-performance-mode")]
 use uuid::Uuid;
 
-use std::collections::HashMap;
+#[cfg(feature = "high-performance-mode")]
+use crate::db::schema::event_log::dsl as event_log;
+use crate::db::{
+    models::ConfigVersion,
+    schema::{
+        config_versions,
+        contexts::dsl::{self as ctxt},
+        default_configs::dsl as def_conf,
+    },
+};
 
 pub fn parse_headermap_safe(headermap: &HeaderMap) -> HashMap<String, String> {
     let mut req_headers = HashMap::new();
