+++ crates/frontend/Cargo.toml
@@ -24,6 +24,7 @@ serde_json = { workspace = true }
 reqwest = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
+superposition_types = { path = "../superposition_types", default-features = false }
 url = "2.5.0"
 wasm-bindgen = "=0.2.89"
 web-sys = { version = "0.3.64", features = [
