+++ crates/frontend/src/types.rs
@@ -263,20 +263,6 @@ impl DropdownOption for DefaultConfig {
     }
 }
 
-#[derive(Deserialize, Serialize, Clone, Debug)]
-pub struct Context {
-    pub id: String,
-    pub condition: Value,
-    pub override_with_keys: [String; 1],
-}
-
-#[derive(Deserialize, Serialize, Clone, Debug, Default)]
-pub struct Config {
-    pub contexts: Vec<Context>,
-    pub overrides: Map<String, Value>,
-    pub default_configs: Map<String, Value>,
-}
-
 pub type FunctionsName = String;
 impl DropdownOption for FunctionsName {
     fn key(&self) -> String {
