+++ crates/superposition_types/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 # env
-actix-web = { workspace = true }
+actix-web = { workspace = true, optional = true }
 log = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
@@ -20,9 +20,11 @@ regex = { workspace = true }
 superposition_derives = { path = "../superposition_derives", optional = true }
 
 [features]
-disable_db_data_validation = []
-result = ["dep:diesel", "dep:anyhow", "dep:thiserror"]
+default = ["server"]
 diesel_derives = ["dep:diesel", "dep:superposition_derives"]
+disable_db_data_validation = []
+result = ["dep:diesel", "dep:anyhow", "dep:thiserror", "dep:actix-web"]
+server = ["dep:actix-web"]
 
 [lints]
 workspace = true
